Cleaned up console display, still need to add some documentation. Our Sql query is acting up again, need to fix

+++ Capstone/Capstone/Models/Spaces.cs
@@ -21,7 +21,7 @@ namespace Capstone.Models
 
         public int SpaceMaxOccupancy { get; set; }
 
-        public int TotalCost { get; set; }
+        public decimal TotalCost { get; set; }
         
     }
 }
